Refactor ref counting in pubsub. Fix ref counting bug.
