+++ pkg/runtime/fx.go
@@ -1,0 +1,8 @@
+package runtime
+
+/************************************************************
+ *                                                          *
+ *  fx.go exports the Wetware runtime to the Fx framework.  *
+ *  See go.uber.org/fx                                      *
+ *                                                          *
+ ************************************************************/
